TQS-80 swagger functional
