+++ backend/src/main/resources/application.properties
@@ -4,3 +4,7 @@ spring.datasource.username=postgres
 spring.datasource.password=password_tqs
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=update
+springdoc.swagger-ui.path=/docs
+
+spring.security.user.name=admin
+spring.security.user.password=beauty_plaza_admin
